Fix compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@18251 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/RepositoryBrowserSelection.cpp
@@ -147,7 +147,7 @@ CRepositoryBrowserSelection::FindURL (const CTSVNPath& url) const
 			if (GetURL (i, k).IsEquivalentTo (url))
 				return std::make_pair (i, k);
 
-	return std::pair<size_t, size_t>(-1, -1);
+	return std::pair<size_t, size_t>((size_t)-1, (size_t)-1);
 }
 
 bool CRepositoryBrowserSelection::Contains (const CTSVNPath& url) const
